requirements.txt
numpy
Pillow
PyYAML>=5.3.1
torch
torchvision
tqdm
argparse